+++ src/Metadata/Resource/Factory/OperationDefaultsTrait.php
@@ -16,6 +16,7 @@ namespace ApiPlatform\Metadata\Resource\Factory;
 use ApiPlatform\Metadata\ApiResource;
 use ApiPlatform\Metadata\CollectionOperationInterface;
 use ApiPlatform\Metadata\Delete;
+use ApiPlatform\Metadata\Error;
 use ApiPlatform\Metadata\Exception\RuntimeException;
 use ApiPlatform\Metadata\Get;
 use ApiPlatform\Metadata\GetCollection;
@@ -32,6 +33,7 @@ use ApiPlatform\Metadata\Patch;
 use ApiPlatform\Metadata\Post;
 use ApiPlatform\Metadata\Util\CamelCaseToSnakeCaseNameConverter;
 use ApiPlatform\State\CreateProvider;
+use ApiPlatform\Validator\Exception\ValidationException;
 use Psr\Log\LoggerInterface;
 
 trait OperationDefaultsTrait
@@ -42,6 +44,11 @@ trait OperationDefaultsTrait
 
     private function addGlobalDefaults(ApiResource|Operation $operation): ApiResource|Operation
     {
+        // Do not add global defaults for internal resources:
+        if (\in_array($operation->getClass(), [Error::class, ValidationException::class], true)) {
+            return $operation;
+        }
+
         $extraProperties = $this->defaults['extra_properties'] ?? [];
 
         foreach ($this->defaults as $key => $value) {